+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 - Ensure BrowserRouter mounts first.
 - Allow an empty string to remove query params in `useNavigate`.
 
+## Next
+
+- Allow setting of query string.
+
 ## 0.2.4
 
 - Mapped history `back`, `forward` and `go` onto `useNavigate`.